IvanR - Added the Main Collection view custom layout, also changed the app to point to pivotal’s new domain and fixed a bug where an contact object can now return an array of emails

+++ Book Room/BRMeetingRoomsCollectionViewController.m
@@ -61,7 +61,7 @@ static NSInteger const kMeetingRoomsFetchStep = 20;
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH, 0), ^{
         NSMutableArray *rooms = [NSMutableArray array];
 
-        for (int i=count;i<roomsCount+kMeetingRoomsFetchStep;i++) {
+        for (NSUInteger i=count;i<roomsCount+kMeetingRoomsFetchStep;i++) {
             if (i >= meetingRooms.count) break;
 
             NSDictionary *room = meetingRooms[i];
